requirements.txt
whisper==1.1.10
pydub==0.25.1
torch==2.1.0
python-dotenv==1.0.0
google-generativeai==0.3.1
pytube==15.0.0
pyttsx3==2.90
yt_dlp==2022.1.1
moviepy==1.0.3

#To use: pip install -r requirements.txt